+++ src/Video/ProcessParams.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Soluble\MediaTools\Video;
+
+use Soluble\MediaTools\Common\Process\ProcessParams as CommonProcessParams;
+
+class ProcessParams extends CommonProcessParams
+{
+}